Moved HeroStatistic and HeroAbility objects to hero module.
