+++ hero/admin.py
@@ -1,12 +1,12 @@
 from django.contrib import admin
 from .models.hero import Hero
 from hero.models.occupation import Occupation
-from hero.models.ability import Ability, HeroAbility
-from hero.models.statistic import Statistic
+from hero.models.ability import Ability
+from hero.models.hero import HeroAbility, HeroStatistic
 
 
 admin.site.register(Hero)
 admin.site.register(Occupation)
 admin.site.register(Ability)
 admin.site.register(HeroAbility)
-admin.site.register(Statistic)
+admin.site.register(HeroStatistic)
