+++ hero_api/urls.py
@@ -8,7 +8,7 @@ app_name = 'hero_api'
 router = routers.DefaultRouter()
 router.register('hero', views.HeroViewSet)
 router.register('hero/(?P<hero_pk>[0-9]+)/hero_ability', views.HeroAbilityViewSet)
-router.register('hero/(?P<hero_pk>[0-9]+)/statistic', views.StatisticViewSet)
+router.register('hero/(?P<hero_pk>[0-9]+)/statistic', views.HeroStatisticViewSet)
 router.register('ability', views.AbilityViewSet)
 router.register('occupation', views.OccupationViewSet)
